feat: setting up cities search screen  - fixing the search bar  - observing the cities list from the view-model  - dispalying the cities list on the the lazy column as DomainEntities

+++ app/src/main/java/com/example/weatherapp/operations/data_management/data_entities/CacheEntity.kt
@@ -21,7 +21,7 @@ data class CacheEntity(
 
     // current.
     val lastUpdated: String? = null,
-    val tempC: Int? = null,
+    val tempC: Double? = null,
     val feelslikeC: Double? = null,
     val condition_txt: String? = null,
     val condition_ic: String? = null,
