+++ app/src/main/java/com/example/weatherapp/operations/data_management/data_entities/NetWorkEntitySubClasses/Current.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Current (
     @SerializedName("last_updated"       ) var lastUpdated      : String?     = null,//
-    @SerializedName("temp_c"             ) var tempC            : Int?        = null,//
+    @SerializedName("temp_c"             ) var tempC            : Double?        = null,//
     @SerializedName("is_day"             ) var isDay            : Int?        = null,
     @SerializedName("condition"          ) var condition        : Condition?  = Condition(),//
     @SerializedName("wind_kph"           ) var windKph          : Double?     = null,//
